backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
Authlib==1.4.0
bcrypt==4.2.1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
dnspython==2.0.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
fuzzywuzzy==0.18.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.9.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
